pyproject.toml
[project]
name = "moodle-mcp"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server implementation that provides capabilities to interact with Moodle LMS."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "glom>=24.11.0",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[project.scripts]
moodle-mcp = "moodle_mcp.server:main"

[project.urls]
Homepage = "https://github.com/loyaniu/moodle-mcp"
